refactor: useCard validate 로직 분리
